iac_test_framework/requirements.txt
boto3==1.34.0
pytest==7.4.0
pyyaml==6.0.1